- Remove bug in app/usercommandeditdialog.cpp which caused changes to a user snippet not to be saved when there is only one snippet. - Load the "Replace", "Go to line" and "Indent" panels when required instead of at startup of the application. - Move the command completion code completely to app/tikzeditor.cpp. - Remove bug which caused the application to crash when command completion is disabled and the settings dialog or the user commands editing dialog are accepted by the user. - Various smaller cleanups.
git-svn-id: file:///tmp/svnroot/ktikz/trunk@203 68211431-5336-0410-8235-f1471971ebd9

+++ common/tikzpreview.cpp
@@ -486,9 +486,9 @@ void TikzPreview::mouseMoveEvent(QMouseEvent *event)
 				invUnitY *= 10;
 		}
 
-		qreal cursorX = event->x() + horizontalScrollBar()->value() - qMax(0.0, viewport()->width() - m_tikzPixmapItem->boundingRect().width()) / 2;
+		qreal cursorX = event->x() + horizontalScrollBar()->value() - qMax(qreal(0.0), viewport()->width() - m_tikzPixmapItem->boundingRect().width()) / 2;
 		cursorX /= m_zoomFactor;
-		qreal cursorY = event->y() + verticalScrollBar()->value() - qMax(0.0, viewport()->height() - m_tikzPixmapItem->boundingRect().height()) / 2;
+		qreal cursorY = event->y() + verticalScrollBar()->value() - qMax(qreal(0.0), viewport()->height() - m_tikzPixmapItem->boundingRect().height()) / 2;
 		cursorY /= m_zoomFactor;
 		const qreal coordX = cursorX + minX;
 		const qreal coordY = maxY - cursorY;
